DO NOT MERGE: Experiments with instruction encoding extraction
The tools in this directory product JSON files describing instruction
encodings based on the binutils overlay file xtensa-modules.c.  The
input format accepted is a .tar.gz or .tar.bz2 overlay tarball as
accepted by a crosstools-ng build.  Just run "make" in this directory
to produce a JSON file for each machine description.  (It takes
surprisingly long to compile these files, so feel free to build in
parallel to speed the process!).

Signed-off-by: Andy Ross <andyross@google.com>